requirements.txt
ansible==2.8.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
Django==2.2.4
entrypoints==0.3
Fabric3==1.14.post1
filelock==3.0.12
flake8==3.7.8
importlib-metadata==0.19
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
paramiko==2.6.0
pep8==1.7.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.0.1
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tox==3.13.2
virtualenv==16.7.2
wcwidth==0.1.7
zipp==0.5.2
